Pass transaction model as environment object
