+++ HomeBugh/Screens/TransactionsView/TransactionsView.swift
@@ -7,24 +7,9 @@
 
 import SwiftUI
 
-class Transactions: ObservableObject {
-    @Published var items = [
-        Transaction(id: 0,
-                    amount: "18,23$",
-                    comment: "Magnesium",
-                    category: Category(id: 0, name: "Food", categoryType: CategoryType(id: 0, name: "Spending"), inactive: false),
-                    account: Account(id: 0, name: "Deutsche Bank", balance: 100000.0, currency: Currency(id: 0, name: "Euro", unit: "Euro"), status: "active", showInSummary: true)),
-        Transaction(id: 1,
-                    amount: "5$",
-                    comment: "",
-                    category: Category(id: 1, name: "Food", categoryType: CategoryType(id: 1, name: "Spending"), inactive: false),
-                    account: Account(id: 1, name: "Bar", balance: 1000.0, currency: Currency(id: 1, name: "Euro", unit: "Euro"), status: "active", showInSummary: true))
-    ]
-}
-
 struct TransactionsView: View {
     
-    @ObservedObject var transactions = Transactions()
+    @EnvironmentObject var transactions: Transactions
     
     @State private var addTransactionViewVisible = false
     @State private var showingAlert = false
